# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Portal Berita",
        "description": "Portal Berita API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "Local server"
        },
        {
            "url": "http://staging.example.com",
            "description": "Staging server"
        },
        {
            "url": "http://example.com",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/api/buku": {
            "get": {
                "tags": [
                    "Buku"
                ],
                "summary": "Menampilkan semua buku",
                "operationId": "bukuIndex",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data buku",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Berhasil mengambil data buku"
                                        },
                                        "buku": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BukuSchema"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Kesalahan saat mengambil data buku",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Terjadi kesalahan saat mengambil data buku"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Buku"
                ],
                "summary": "Menambahkan buku baru",
                "operationId": "bukuStore",
                "requestBody": {
                    "description": "Data buku yang akan ditambahkan",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "kategori_id",
                                    "nama_buku",
                                    "judul",
                                    "penulis",
                                    "penerbit",
                                    "jumlah_tersedia"
                                ],
                                "properties": {
                                    "kategori_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nama_buku": {
                                        "type": "string",
                                        "example": "Harry Potter"
                                    },
                                    "judul": {
                                        "type": "string",
                                        "example": "Harry Potter and the Philosopher's Stone"
                                    },
                                    "penulis": {
                                        "type": "string",
                                        "example": "J.K. Rowling"
                                    },
                                    "penerbit": {
                                        "type": "string",
                                        "example": "Gramedia"
                                    },
                                    "tahun_penerbitan": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2001-01-01"
                                    },
                                    "jumlah_tersedia": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Buku berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Buku berhasil ditambahkan"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/BukuSchema"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validasi gagal"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "nama_buku": [
                                                    "Nama buku harus diisi"
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Kesalahan server",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Terjadi kesalahan"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/buku/kategori-list": {
            "get": {
                "tags": [
                    "Buku"
                ],
                "summary": "Menampilkan semua kategori buku",
                "operationId": "bukuKategoriList",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data kategori",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Berhasil mengambil data kategori"
                                        },
                                        "kategori": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/KategoriBukuSchema"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Kesalahan saat mengambil data kategori",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Terjadi kesalahan saat mengambil data kategori"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "KategoriBukuSchema": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nama_kategori": {
                        "type": "string",
                        "example": "Fiksi"
                    },
                    "deskripsi": {
                        "type": "string",
                        "example": "Buku-buku fiksi dan novel"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}